requirements.txt
discord.py-self
colorama
curl_cffi
httpx
asyncio
python-dotenv
pyYAML
requests
openai
fastapi
uvicorn
jinja2